Fix incorrect number of items in custom recipes

+++ src/main/java/mod/maxbogomol/wizards_reborn/common/spell/ray/EarthRaySpell.java
@@ -2,6 +2,7 @@ package mod.maxbogomol.wizards_reborn.common.spell.ray;
 
 import mod.maxbogomol.wizards_reborn.WizardsReborn;
 import mod.maxbogomol.wizards_reborn.api.crystal.CrystalUtils;
+import mod.maxbogomol.wizards_reborn.api.wissen.WissenItemUtils;
 import mod.maxbogomol.wizards_reborn.common.entity.SpellProjectileEntity;
 import mod.maxbogomol.wizards_reborn.common.network.PacketHandler;
 import mod.maxbogomol.wizards_reborn.common.network.spell.EarthRaySpellEffectPacket;
@@ -34,14 +35,18 @@ public class EarthRaySpell extends RaySpell {
     public void onImpact(HitResult ray, Level world, SpellProjectileEntity projectile, Player player, Entity target) {
         super.onImpact(ray, world, projectile, player, target);
 
-        if (target.tickCount % 10 == 0) {
-            ItemStack stack = player.getItemInHand(player.getUsedItemHand());
-            removeWissen(stack, projectile.getStats());
-            int focusLevel = CrystalUtils.getStatLevel(projectile.getStats(), WizardsReborn.FOCUS_CRYSTAL_STAT);
-            float damage = (float) (2.0f + (focusLevel * 0.5));
-            DamageSource damageSource = new DamageSource(target.damageSources().generic().typeHolder(), projectile, player);
-            target.hurt(new DamageSource(target.damageSources().generic().typeHolder(), projectile, player), damage);
-            player.getInventory().hurtArmor(damageSource, damage, Inventory.ALL_ARMOR_SLOTS);
+        if (player != null) {
+            if (target.tickCount % 10 == 0) {
+                ItemStack stack = player.getItemInHand(player.getUsedItemHand());
+                if (WissenItemUtils.canRemoveWissen(stack, getWissenCostWithStat(projectile.getStats(), player))) {
+                    removeWissen(stack, projectile.getStats(), player);
+                    int focusLevel = CrystalUtils.getStatLevel(projectile.getStats(), WizardsReborn.FOCUS_CRYSTAL_STAT);
+                    float damage = (float) (2.0f + (focusLevel * 0.5));
+                    DamageSource damageSource = new DamageSource(target.damageSources().generic().typeHolder(), projectile, player);
+                    target.hurt(new DamageSource(target.damageSources().generic().typeHolder(), projectile, player), damage);
+                    player.getInventory().hurtArmor(damageSource, damage, Inventory.ALL_ARMOR_SLOTS);
+                }
+            }
         }
     }
 
@@ -52,23 +57,25 @@ public class EarthRaySpell extends RaySpell {
         if (player != null) {
             if (player.isShiftKeyDown()) {
                 int focusLevel = CrystalUtils.getStatLevel(projectile.getStats(), WizardsReborn.FOCUS_CRYSTAL_STAT);
+                ItemStack stack = player.getItemInHand(player.getUsedItemHand());
                 if (projectile.tickCount % (20 - (focusLevel * 3)) == 0) {
-                    Vec3 vec = getBlockHitOffset(ray, projectile, 0.1f);
-                    BlockPos blockPos = new BlockPos((int) vec.x(), (int) vec.y(), (int) vec.z());
-                    BlockState blockState = world.getBlockState(blockPos);
-                    if (!blockState.isAir()) {
-                        if (canBreak(blockState)) {
-                            world.destroyBlock(blockPos, true);
+                    if (WissenItemUtils.canRemoveWissen(stack, getWissenCostWithStat(projectile.getStats(), player))) {
+                        Vec3 vec = getBlockHitOffset(ray, projectile, 0.1f);
+                        BlockPos blockPos = new BlockPos((int) vec.x(), (int) vec.y(), (int) vec.z());
+                        BlockState blockState = world.getBlockState(blockPos);
+                        if (!blockState.isAir()) {
+                            if (canBreak(blockState)) {
+                                world.destroyBlock(blockPos, true);
 
-                            ItemStack stack = player.getItemInHand(player.getUsedItemHand());
-                            removeWissen(stack, projectile.getStats());
+                                removeWissen(stack, projectile.getStats(), player);
 
-                            Color color = getColor();
-                            float r = color.getRed() / 255f;
-                            float g = color.getGreen() / 255f;
-                            float b = color.getBlue() / 255f;
+                                Color color = getColor();
+                                float r = color.getRed() / 255f;
+                                float g = color.getGreen() / 255f;
+                                float b = color.getBlue() / 255f;
 
-                            PacketHandler.sendToTracking(world, player.getOnPos(), new EarthRaySpellEffectPacket((float) blockPos.getX() + 0.5f, (float) blockPos.getY() + 0.5f, (float) blockPos.getZ() + 0.5f, r, g, b));
+                                PacketHandler.sendToTracking(world, player.getOnPos(), new EarthRaySpellEffectPacket((float) blockPos.getX() + 0.5f, (float) blockPos.getY() + 0.5f, (float) blockPos.getZ() + 0.5f, r, g, b));
+                            }
                         }
                     }
                 }
